// src/app/shared/enums/portal-placement.ts
export enum PortalPlacement {
  TopCenter = 'top-center',
  BottomCenter = 'bottom-center',
  LeftCenter = 'start-center',
  RightCenter = 'end-center',
  RightTop = 'end-top',
  BottomStart = 'bottom-start',
  BottomStartHorizontal = 'bottom-start-horizontal',
  BottomEnd = 'bottom-end',
  StartCenter = 'start-center',
  StartEnd = 'start-end',
  TopStart = 'top-start',
  TopEnd = 'top-end',
}